Fix sign extension bug for EQU constants >= 0x8000
Problem: EQU constants with values 0x8000..0xFFFF were stored as
int16_t in the object file header, causing sign extension (0x8000
became 0xFFFF8000) when read back by the linker as int. This resulted
in incorrect symbol display ($FFFF8000 instead of $8000) and
potentially wrong results in arithmetic operations like SHR.

Solution: Added REC_CONST_CUSTOM_UNSIGNED record type (following the
same pattern as the existing REC_EXPR_PUSH_VALUE_UNSIGNED) so the
linker knows to read the value as uint16_t instead of int16_t.

Changes:
- include/obj.h: Add REC_CONST_CUSTOM_UNSIGNED enum value
- hcasm/parser.c: Use REC_CONST_CUSTOM_UNSIGNED when e->value >= INT16_MAX
- hcasm/out.c: Add dump case for REC_CONST_CUSTOM_UNSIGNED
- hclink/process.c: Handle REC_CONST_CUSTOM_UNSIGNED with (uint16_t) cast
- hclink/consts.c: Cast to (uint16_t) in fprintf for correct display
- hclink/debug.c: Cast to (uint16_t) in fprintf for correct display

Before: MY_PORT equ 0x8000 -> symbol file shows $FFFF8000
After:  MY_PORT equ 0x8000 -> symbol file shows $8000

+++ hclink/consts.c
@@ -284,7 +284,7 @@ void consts_print(char *sym_name)
     {
         if (c->is_global && c->obj->use_in_link)
         {
-            fprintf(sym, "$%04X %s [GLOBAL %s %s]\n", c->value, c->name, c->is_offset ? "LABEL" : "CONST", c->obj->name);
+            fprintf(sym, "$%04X %s [GLOBAL %s %s]\n", (uint16_t)c->value, c->name, c->is_offset ? "LABEL" : "CONST", c->obj->name);
         }
         c = c->next;
     }
@@ -295,7 +295,7 @@ void consts_print(char *sym_name)
     {
         if (!c->is_global && c->obj->use_in_link)
         {
-            fprintf(sym, "$%04X %s [LOCAL %s %s]\n", c->value, c->name, c->is_offset ? "LABEL" : "CONST", c->obj->name);
+            fprintf(sym, "$%04X %s [LOCAL %s %s]\n", (uint16_t)c->value, c->name, c->is_offset ? "LABEL" : "CONST", c->obj->name);
         }
         c = c->next;
     }
